pull

+++ includes/payments/class-getpaid-payment-form.php
@@ -286,7 +286,7 @@ class GetPaid_Payment_Form extends GetPaid_Data {
 
 				if ( $item->can_purchase() ) {
 
-					$value = (int) $value;
+					$value = (float) $value;
 					$item->set_quantity( $value );
 					if ( 0 == $value ) {
 						$item->set_quantity( 1 );
@@ -416,7 +416,7 @@ class GetPaid_Payment_Form extends GetPaid_Data {
 	 *
 	 * @since 1.0.19
 	 * @param  string $context View or edit context.
-	 * @return array
+	 * @return float
 	 */
 	public function get_earned( $context = 'view' ) {
 		return $this->get_prop( 'earned', $context );
@@ -427,7 +427,7 @@ class GetPaid_Payment_Form extends GetPaid_Data {
 	 *
 	 * @since 1.0.19
 	 * @param  string $context View or edit context.
-	 * @return array
+	 * @return float
 	 */
 	public function get_refunded( $context = 'view' ) {
 		return $this->get_prop( 'refunded', $context );
@@ -438,7 +438,7 @@ class GetPaid_Payment_Form extends GetPaid_Data {
 	 *
 	 * @since 1.0.19
 	 * @param  string $context View or edit context.
-	 * @return array
+	 * @return float
 	 */
 	public function get_cancelled( $context = 'view' ) {
 		return $this->get_prop( 'cancelled', $context );
@@ -449,7 +449,7 @@ class GetPaid_Payment_Form extends GetPaid_Data {
 	 *
 	 * @since 1.0.19
 	 * @param  string $context View or edit context.
-	 * @return array
+	 * @return float
 	 */
 	public function get_failed( $context = 'view' ) {
 		return $this->get_prop( 'failed', $context );
